branch map_2.0 -Markers being pulled from the db

+++ FoodStand/db/migrate/20171026143323_create_beverages.rb
@@ -2,7 +2,7 @@ class CreateBeverages < ActiveRecord::Migration[5.1]
   def change
     create_table :beverages do |t|
       t.string :typesOfDrinks
-      t.float :price
+      t.integer :price
 
       t.timestamps
     end
